Manage LVM locks
It's not possible de take 2 snapshot of the same LVM volume at the same time.
This is a problem when using a single file system (backed by a single LVM volume) to store
several VM as trying to run 2 backups at the same time will fail (at least, taking the snapshot for one will fail).

Add locking before taking a snapshot and wait up to 10 sec for the lock to be released. This should fix
concurrent backups